indexer test is running but it fails
