+++ testnet/stacks-node/src/tests/l1_observer_test.rs
@@ -1,90 +1,19 @@
 use std;
-use std::process::{Child, Command, Stdio};
 use std::thread;
 
 use crate::neon;
+use crate::tests::StacksL1Controller;
+use clarity::util::hash::to_hex;
+use rand::RngCore;
 use stacks::burnchains::Burnchain;
 use std::env;
-use std::io::{BufRead, BufReader};
 use std::time::Duration;
 
-#[derive(std::fmt::Debug)]
-pub enum SubprocessError {
-    SpawnFailed(String),
-}
-
-type SubprocessResult<T> = Result<T, SubprocessError>;
-
-/// In charge of running L1 `stacks-node`.
-pub struct StacksL1Controller {
-    sub_process: Option<Child>,
-    config_path: String,
-    out_reader: Option<BufReader<std::process::ChildStdout>>,
-}
-
-impl StacksL1Controller {
-    pub fn new(config_path: String) -> StacksL1Controller {
-        StacksL1Controller {
-            sub_process: None,
-            config_path,
-            out_reader: None,
-        }
-    }
-
-    pub fn start_process(&mut self) -> SubprocessResult<()> {
-        let binary = match env::var("STACKS_BASE_DIR") {
-            Err(_) => {
-                // assume stacks-node is in path
-                "stacks-node".into()
-            }
-            Ok(path) => path,
-        };
-        let mut command = Command::new(&binary);
-        command
-            .stdout(Stdio::piped())
-            .arg("start")
-            .arg("--config=".to_owned() + &self.config_path);
-
-        info!("stacks-node mainchain spawn: {:?}", command);
-
-        let mut process = match command.spawn() {
-            Ok(child) => child,
-            Err(e) => return Err(SubprocessError::SpawnFailed(format!("{:?}", e))),
-        };
-
-        info!("stacks-node mainchain spawned, waiting for startup");
-        let mut out_reader = BufReader::new(process.stdout.take().unwrap());
-
-        let mut line = String::new();
-        while let Ok(bytes_read) = out_reader.read_line(&mut line) {
-            if bytes_read == 0 {
-                return Err(SubprocessError::SpawnFailed(
-                    "Stacks L1 closed before spawning network".into(),
-                ));
-            }
-            info!("{:?}", &line);
-            break;
-        }
-
-        info!("Stacks L1 startup finished");
-
-        self.sub_process = Some(process);
-        self.out_reader = Some(out_reader);
-
-        Ok(())
-    }
-
-    pub fn kill_process(&mut self) {
-        if let Some(mut sub_process) = self.sub_process.take() {
-            sub_process.kill().unwrap();
-        }
-    }
-}
-
-impl Drop for StacksL1Controller {
-    fn drop(&mut self) {
-        self.kill_process();
-    }
+fn random_sortdb_test_dir() -> String {
+    let mut rng = rand::thread_rng();
+    let mut buf = [0u8; 32];
+    rng.fill_bytes(&mut buf);
+    format!("/tmp/stacks-node-tests/sortdb/test-{}", to_hex(&buf))
 }
 
 /// This test brings up the Stacks-L1 chain in "mocknet" mode, and ensures that our listener can hear and record burn blocks
@@ -106,6 +35,12 @@ fn l1_observer_test() {
     let mut config = super::new_test_conf();
     config.burnchain.chain = "stacks_layer_1".to_string();
     config.burnchain.mode = "hyperchain".to_string();
+
+    let db_path_dir = random_sortdb_test_dir();
+    config.burnchain.indexer_base_db_path = db_path_dir;
+    config.burnchain.first_burn_header_hash =
+        "1111111111111111111111111111111111111111111111111111111111111111".to_string();
+
     let mut run_loop = neon::RunLoop::new(config.clone());
     let channel = run_loop.get_coordinator_channel().unwrap();
     thread::spawn(move || run_loop.start(None, 0));
